python_api/requirements.txt
fastapi
google-cloud-bigquery
db-dtypes